Extract document_view load fns into agent/document_view/load.rs
